requirements.txt
oauthlib==0.7.2
python-decouple==2.3
requests==2.7.0
requests-oauthlib==0.5.0
six==1.9.0
tweepy==3.3.0
